refactor(game): split game footer
