+++ pages/game/[gameId].tsx
@@ -6,7 +6,7 @@ import { useSocket } from '../../utils/SocketProvider';
 import Container from '../../components/columnLayout';
 import Header from '../../components/game/gameHeader';
 import GameBackGround from '../../components/game/background';
-import UserInfo from '../../components/userInfo';
+import GameFooter from '../../components/game/gameFooter';
 
 interface UserData {
   id: number;
@@ -35,6 +35,7 @@ const GamePage: React.FC = () => {
     intraName: '',
     avatar: '',
   };
+  const [countdown, setCountdown] = useState<number>(0);
 
   // canvas
   const canvasRef = useRef<HTMLCanvasElement>(null);
@@ -90,6 +91,15 @@ const GamePage: React.FC = () => {
 
       const handleGameStart = () => {
         console.log('handleGameStart');
+        setCountdown(3);
+        const countdownInterval = setInterval(() => {
+          setCountdown((prevCountdown) => prevCountdown - 1);
+        }, 1000);
+
+        setTimeout(() => {
+          clearInterval(countdownInterval);
+          setCountdown(0);
+        }, 3000);
       };
 
       const handleKeyDown = (event: KeyboardEvent) => {
@@ -220,117 +230,18 @@ const GamePage: React.FC = () => {
     }
   };
 
-  const handleAvatarPath = (avatar: string) => {
-    const apiUrl = 'http://localhost:8080/';
-    return apiUrl + avatar;
-  };
-
   return (
     <Container>
       <Header title='General Game' text='일반 게임' />
       <GameBackGround width={width} height={height} canvasRef={canvasRef} />
-      <GameContent>
-        <UserInfoFrame sort='flex-start'>
-          <UserImage
-            src={handleAvatarPath(leftUser.avatar)}
-            alt='Uploaded Image'
-            width={300}
-            height={300}
-          />
-          <UserTextFrame sort='flex-start'>
-            <Text textsize='1.1vw'> {leftUser.nickName} </Text>
-            <Text textsize='0.7vw'> {leftUser.intraName} </Text>
-          </UserTextFrame>
-        </UserInfoFrame>
-        <ScoreBoardFrame>
-          <ScoreBoard color='1'> {leftScore} </ScoreBoard>
-          <ScoreBoard color='2'> {rightScore} </ScoreBoard>
-        </ScoreBoardFrame>
-        <UserInfoFrame sort='flex-end'>
-          <UserTextFrame sort='flex-end'>
-            <Text textsize='1.1vw'> {rightUser.nickName} </Text>
-            <Text textsize='0.7vw'> {rightUser.intraName} </Text>
-          </UserTextFrame>
-          <UserImage
-            src={handleAvatarPath(rightUser.avatar)}
-            alt='Uploaded Image'
-            width={300}
-            height={300}
-          />
-        </UserInfoFrame>
-      </GameContent>
+      <GameFooter
+        leftUser={leftUser}
+        rightUser={rightUser}
+        leftScore={leftScore}
+        rightScore={rightScore}
+      />
     </Container>
   );
 };
 
 export default GamePage;
-
-const GameContent = styled.div`
-  width: 70%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 1%;
-`;
-
-const ScoreBoardFrame = styled.div`
-  width: 20%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-`;
-
-const ScoreBoard = styled.div<{ color: string }>`
-  width: 40%;
-  height: 100%;
-  background-color: ${(props) => {
-    switch (props.color) {
-      case '1':
-        return props.theme.colors.blue;
-      case '2':
-        return props.theme.colors.red;
-    }
-  }};
-  color: ${(props) => props.theme.colors.cream};
-  font-size: 3vh;
-  font-family: 'GiantsLight';
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-`;
-
-const UserInfoFrame = styled.div<{ sort: string }>`
-  width: 20%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: ${(props) => props.sort};
-  gap: 1vw;
-`;
-
-const UserImage = styled(Image)`
-  width: 3vw;
-  height: auto;
-  border-radius: 50px;
-`;
-
-const UserTextFrame = styled.div<{ sort: string }>`
-  display: flex;
-  flex-direction: column;
-  align-items: ${(props) => props.sort};
-  justify-content: center;
-  gap: 0.5vw;
-`;
-
-const Text = styled.div<{ textsize: string }>`
-  white-space: normal;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-family: 'GiantsLight';
-  color: ${(props) => props.theme.colors.brown};
-  font-size: ${(props) => props.textsize};
-`;
